requirements.txt
tensorflow-gpu>=2.0
tensorflow-addons
pep8
pylint
mypy